+++ crates/api/src/systems/event_state_system.rs
@@ -98,9 +98,8 @@ impl EventStateSystem {
 
         let theme = ecm
             .component_store()
-            .get::<Global>("global", root)
+            .get::<Theme>("theme", root)
             .unwrap()
-            .theme
             .clone();
 
         // global key handling
@@ -388,9 +387,8 @@ impl System<Tree, StringComponentStore, RenderContext2D> for EventStateSystem {
 
             let theme = ecm
                 .component_store()
-                .get::<Global>("global", root)
+                .get::<Theme>("theme", root)
                 .unwrap()
-                .theme
                 .clone();
 
             let mut remove_widget_list: Vec<Entity> = vec![];