+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Raise on_changed callback also on shared widgets
 * Localization
 * Custom theme names for types Brush, String, Thickness, f32, f64
+* Add focus_state property to Window
+* Add theme property to Window
 
 ### 0.3.1-alpha3